requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field[phonenumberslite]==7.3.*
djangorestframework==3.14.*
markdown==3.5.*
django-filter==23.5
pydantic==2.5.*
